修复安卓端编译 后台管理增加操作日志 怪物装备掉落概率支持 0.01% 到 100% 可以自定义场景刷怪 优化脱离卡死的位置的选择逻辑 修复传送石 死亡复活位置调整到最近安全区域

+++ server/mysql/migrate_add_monster_multi_drops.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS `server_monster_drops` (
   `monster_id` SMALLINT UNSIGNED NOT NULL,
   `drop_slot` TINYINT UNSIGNED NOT NULL,
   `item_id` INT UNSIGNED NOT NULL,
-  `drop_rate_percent` TINYINT UNSIGNED NOT NULL,
+  `drop_rate_percent` DECIMAL(5,2) UNSIGNED NOT NULL,
   PRIMARY KEY (`monster_id`, `drop_slot`),
   CONSTRAINT `fk_server_monster_drops_monster`
     FOREIGN KEY (`monster_id`) REFERENCES `server_monsters` (`monster_id`)
